TypeSupport correctly supports Flamingo.idl
Proper FlamingoTypeSupport.

Previously (if you look in the original FLamingoTypeSupport.idl) the compiler was not identifying the struct as a topic type, because there was no @topic declaration before the struct. Make sure when you use tao/dds compilers that you set --idl-version to 4, and it must be ran in the main directory to properly grab any include files.